Add notion of forwarded APIs to the catalog
A very common problem when analyzing missing APIs are false positives for APIs
on TypeInfo.

They result from our creative work in System.Reflection where .NET Core 1
(and .NET Standard 1.x) had System.Object separated from reflection by the
introduction of TypeInfo. In the reference assembly, Type and TypeInfo were
unrelated in the hierarchy.

In .NET Core 2 (and .NET Standard 2) we made reflection compatible with .NET
Framework by making TypeInfo extend Type. In that world, TypeInfo is
basically empty because all the reflection APIs are inherited from Type.

Code compiled against .NET Core 1.x/.NET Standard 1.x refers to reflection
APIs via TypeInfo but they no longer live there, they already live on the
base. In IL that's not a problem because it's valid to refer to a member via
its derived type. But the catalog doesn't know that because going from a
derived member to a base member is actually fairly involved and requires
matching signatures.

We could decide to solve this by adding a way to generally lookup base members
but that feels over engineered. Instead I added the notion of forwarded APIs
which I currently just hard wire to forward all members of TypeInfo on Type,
when their relative names match. That's not a great general solution but
works well enough for reflection.

This fixes #23.