Forward Type members to MemberInfo
With 5ec19504 we added the notion of forwarded API in order to deal with the
creative work we did with System.Type and reflection in .NET Core 1.0.

However, there are three types involved:

    TypeInfo --|> Type --|> MemberInfo

(TypeInfo extends Type, which in turn extends MemberInfo).

In 5ec19504 we forwarded members that exist on both TypeInfo and Type to Type.
However, we also need to do the same for Type and MemberInfo.