Provide a built-in mechanism to find APIs by GUID
The dictionary from GUID to ApiModel isn't very compact because the structure
contains a reference to the ApiCatalogModel and the offset. This makes the
dictionary bloated with tons of redundant references. We could store the
offsets but this feels ugly. It seems easier and more convenient to have a
lazily initialized dictionary that is stored in the catalog model and provide
a public method that given a GUID gives you back the ApiModel.