Delete reference units without usages before aggregation
This will exclude all packages that failed processing as well as packages
that don't contain any managed assets from contributing to the overall
percentage.

We know about 1/3 of all packages are in this camp and this significantly
skews the results when trying to compare NuGet to, say, NetFxCompatLab or
Upgrade Planner, where each bucket contains 100% managed assets.