Add GenUsage and GenDesignNotes
Instead of generating usage and design notes when the catalog is re-index,
this splits the work into three workflows:

1. gen-catalog
2. gen-design-notes
3. gen-usage

gen-catalog runs daily but only does the catalog and suffix tree.

gen-design-notes runs twice a weeek after our review sessions uploads the
binary blob.

gen-usage runs every six hours and uses jobs to for both Planner and NuGet.
There is a dependent job that aggregates the result into the single usage
blob that is then uploaded.

Each of those no has a separate web hook URL and notifies the site that new
data is available.