Include version information, for now as a comment
We need to figure out how we want to represent this for indexing and storage.
My gut is that the current storage mechanism in API Catalog is sufficient
because assemblies are de-duplicated. The additional ~350 frameworks cost us
a a few strings for the framework names for each framework an array with
pointers to assemblies. That feels cheap, but we should measure it.

The real questions are:

1. How we want to represent this data in the FrameworkDefinition
2. How to ensure we don't index the same files over and over again during
   catalog construction