Improve storage needs for indexer
Instead of storing all assemblies inline with framework and package documents,
let's store the assemblies as files, named by fingerprint. This makes their
content shareable and reduces the needs in the highly-duplicated
platform-specific frameworks where the vast majority of APIs are shared.

This also moves the APIs into the assembly documents.